refactor: split OpenCode artifact helpers
